convert `UnstableRateCounter` to an abstract class and move triangles implementation to its own component
